Net rs

+++ reactor-core/src/main/java/reactor/core/dispatch/SingleThreadDispatcher.java
@@ -46,7 +46,7 @@ public abstract class SingleThreadDispatcher extends AbstractLifecycleDispatcher
 	}
 
 	@Override
-	public int backlogSize() {
+	public long backlogSize() {
 		return backlog;
 	}
